+++ src/application/middlewares/validate-request.ts
@@ -1,5 +1,6 @@
 import { Request, Response, NextFunction } from 'express'
 import { validationResult } from 'express-validator'
+import { logger } from '../../infra/logger/winston-config-stream'
 import { RequestValidationError } from '../errors/request-validation-error'
 
 export const validateRequest = (
@@ -9,6 +10,7 @@ export const validateRequest = (
 ) => {
   const errors = validationResult(req)
   if (!errors.isEmpty()) {
+    errors.array().forEach((err: any) => logger.error(`${err.msg}`))
     throw new RequestValidationError(errors.array())
   }
   next()