+++ src/domain/game/game-entity.ts
@@ -1,0 +1,17 @@
+import { GameStatus } from '../../application/enums/gameStatus'
+
+export interface Score {
+  correctColors: number
+  correctPositions: number
+}
+
+export interface GameEntity {
+  id: string
+  createdAt: Date
+  state: GameStatus
+  totalGuesses: number
+  guessesRemaining: number
+  board: number[]
+  history: Score[][]
+  userId: string
+}